--- a/README.md
+++ b/README.md
@@ -53,7 +53,7 @@ The following features are implemented:
 | Geom              | `PLANE`, `SPHERE`, `CAPSULE`       |
 | Constraint        | `LIMIT_JOINT`, `CONTACT_PYRAMIDAL` |
 |                   | `CONTACT_ELLIPTIC`                 |
-| Equality          | `JOINT`                            |
+| Equality          | `CONNECT`, `JOINT`                 |
 | Integrator        | `EULER`, `IMPLICITFAST`, `RK4`     |
 | Cone              | `PYRAMIDAL`, `ELLIPTIC`            |
 | Condim            | (1, 3, 4, 6), (1, 3)               |