--- a/mujoco_warp/_src/collision_primitive.py
+++ b/mujoco_warp/_src/collision_primitive.py
@@ -23,6 +23,7 @@ from .types import Data
 from .types import GeomType
 from .types import Model
 
+wp.set_module_options({"enable_backward": False})
 
 @wp.struct
 class Geom: