--- a/mujoco_warp/_src/collision_primitive.py
+++ b/mujoco_warp/_src/collision_primitive.py
@@ -25,6 +25,7 @@ from .types import Model
 
 wp.set_module_options({"enable_backward": False})
 
+
 @wp.struct
 class Geom:
   pos: wp.vec3