--- a/mujoco_warp/_src/support.py
+++ b/mujoco_warp/_src/support.py
@@ -28,6 +28,7 @@ from .warp_util import kernel
 
 wp.set_module_options({"enable_backward": False})
 
+
 def is_sparse(m: mujoco.MjModel):
   if m.opt.jacobian == mujoco.mjtJacobian.mjJAC_AUTO:
     return m.nv >= 60