--- a/mujoco_warp/test_data/constraints.xml
+++ b/mujoco_warp/test_data/constraints.xml
@@ -78,6 +78,8 @@
   </worldbody>
 
   <equality>
+    <connect name="c_site"   site1="site0"      site2="site1"/>
+    <connect name="connect"  body1="anchor1"    body2="beam1"   anchor="1 0 -1" />
     <joint   name="joint"    joint1="joint3"    joint2="joint4" polycoef="0.5 -1 0.1 0.15 0.2" />
   </equality>