~~**WIP**. Passes all tests, but I'm still working on making it faster.~~

#### Before; no connect equalities
(added `wp.set_module_options({"enable_backward": False})` to collision_driver.py, forward.py, collision_primitive.py, support.py so that testspeed runs)
```
mjwarp-testspeed --function=step --mjcf=test_data/humanoid/humanoid.xml --batch_size=8192 --event_trace=True

Summary for 8192 parallel rollouts

 Total JIT time: 5.36 s
 Total simulation time: 9.79 s
 Total steps per second: 836,892
 Total realtime factor: 4,184.46 x
 Total time per step: 1194.90 ns

Event trace:

step: 1193.22
  forward: 1189.60
    fwd_position: 206.42
      kinematics: 46.02
      com_pos: 14.79
      camlight: 3.36
      crb: 44.21
      factor_m: 28.98
      collision: 24.50
      make_constraint: 35.97
      transmission: 7.36
    sensor_pos: 0.14
    fwd_velocity: 53.21
      com_vel: 19.66
      passive: 1.76
      rne: 19.70
    sensor_vel: 0.13
    fwd_actuation: 14.99
    fwd_acceleration: 27.49
      xfrc_accumulate: 9.02
      solve_m: 16.56
    sensor_acc: 0.14
    solve: 885.80
      mul_m: 15.50
      _linesearch: [ 74.45, 43.45, 30.16, 27.18, 25.79, 25.14, 24.74, 24.44, 24.22, 24.14 ]
        mul_m: [ 15.75, 3.36, 1.86, 1.66, 1.64, 1.64, 1.65, 1.65, 1.65, 1.65 ]
  euler: 3.22
```
#### After; no connect equalities
```
mjwarp-testspeed --function=step --
mjcf=test_data/humanoid/humanoid.xml --batch_size=8192 --event_trace=True

Summary for 8192 parallel rollouts

 Total JIT time: 5.70 s
 Total simulation time: 9.74 s
 Total steps per second: 841,284
 Total realtime factor: 4,206.42 x
 Total time per step: 1188.66 ns

Event trace:

step: 1187.00
  forward: 1183.38
    fwd_position: 199.34
      kinematics: 46.21
      com_pos: 14.82
      camlight: 3.32
      crb: 44.26
      factor_m: 29.01
      collision: 24.52
      make_constraint: 28.63
      transmission: 7.35
    sensor_pos: 0.13
    fwd_velocity: 53.31
      com_vel: 19.70
      passive: 1.76
      rne: 19.71
    sensor_vel: 0.14
    fwd_actuation: 15.02
    fwd_acceleration: 27.54
      xfrc_accumulate: 9.05
      solve_m: 16.57
    sensor_acc: 0.14
    solve: 886.48
      mul_m: 15.49
      _linesearch: [ 74.39, 43.43, 30.14, 27.19, 25.74, 25.11, 24.74, 24.43, 24.23, 24.14 ]
        mul_m: [ 15.72, 3.33, 1.83, 1.63, 1.60, 1.61, 1.61, 1.61, 1.60, 1.61 ]
  euler: 3.22
```

#### After, with connect equalities
Add to humanoid.xml:
```xml
  <equality>
    <connect name="joint1" body1="hand_right" body2="hand_left" anchor="0 0 0"/>
    <connect name="joint2" body1="foot_right" body2="foot_left" anchor="0 0 0"/>
  </equality>
```

```
mjwarp-testspeed --function=step --
mjcf=test_data/humanoid/humanoid.xml --batch_size=8192 --event_trace=True

Summary for 8192 parallel rollouts

 Total JIT time: 2.71 s
 Total simulation time: 16.18 s
 Total steps per second: 506,404
 Total realtime factor: 2,532.02 x
 Total time per step: 1974.71 ns

Event trace:

step: 1973.03
  forward: 1969.35
    fwd_position: 218.87
      kinematics: 46.76
      com_pos: 14.88
      camlight: 3.34
      crb: 44.20
      factor_m: 29.00
      collision: 24.50
      make_constraint: 46.88
      transmission: 8.08
    sensor_pos: 0.14
    fwd_velocity: 53.04
      com_vel: 19.50
      passive: 1.74
      rne: 19.69
    sensor_vel: 0.14
    fwd_actuation: 14.95
    fwd_acceleration: 28.06
      xfrc_accumulate: 9.22
      solve_m: 16.87
    sensor_acc: 0.14
    solve: 1652.73
      mul_m: 15.53
      _linesearch: [ 95.88, 84.73, 68.61, 54.28, 45.96, 42.42, 40.87, 40.00, 39.41, 39.06 ]
        mul_m: [ 17.50, 14.70, 7.71, 4.36, 3.37, 3.10, 2.99, 2.96, 2.94, 2.92 ]
  euler: 3.28
```