@Andrew-Luo1 2025-04-11
As seen in the testspeed results, the current implementation is quite slow. The culprit is
```python
      d.efc.J[efcid, dofid] = j1mj2[0]
      d.efc.J[efcid + 1, dofid] = j1mj2[1]
      d.efc.J[efcid + 2, dofid] = j1mj2[2]
```
inside the `m.nv` for loop. I believe that significant speedups are possible from splitting the d.efc.J writing, Jqvel calculation and efc row updating into separate kernels but this would make the code much messier and potentially require additional struct fields.

Let me know if higher performance is a priority, otherwise I think we can go with how it's currently written.