@thowell 2025-04-12
performance improvements can be a future pr or prs. will merge the current code. thanks!